pyproject.toml
[project]
name = "darkforge-memory-battle"
version = "0.1.0"
description = "Reproducible benchmark harness for six agent-memory systems on Dark Forge workloads"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=0.5.20",
    "ollama>=0.4.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.6.0", "pytest>=8.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/darkforge_memory_battle"]

[tool.ruff]
line-length = 100
target-version = "py312"
